requirements.txt
torch
pygame-ce
numpy
matplotlib
